refactor: add SponsorSection
