+++ apps/animated-landing/src/components/sections/SponsorSection.tsx
@@ -1,0 +1,54 @@
+import Image, { StaticImageData } from 'next/image'
+
+import blockStudio from '@/../public/assets/company/block-studio.png'
+import kdanMobile from '@/../public/assets/company/kdan-mobile.png'
+import titanSoft from '@/../public/assets/company/titan-soft.png'
+
+import ExternalLink from '../ExternalLink'
+import YellowStarIcon from '../icons/YellowStarIcon'
+
+type Company = {
+  name: string
+  logo: StaticImageData
+  href: string
+}
+
+const COMPANIES: Company[] = [
+  {
+    name: '板塊設計',
+    logo: blockStudio,
+    href: 'https://blockstudio.tw/',
+  },
+  {
+    name: '凱鈿行動科技',
+    logo: kdanMobile,
+    href: 'https://www.kdanmobile.com/',
+  },
+  {
+    name: '新加坡商 泰坦科技',
+    logo: titanSoft,
+    href: 'https://titansoft.com',
+  },
+]
+
+export default function SponsorSection() {
+  return (
+    <section className="bg-n3 section gap-6 py-10">
+      <div className="relative mb-4 px-6">
+        <YellowStarIcon className="absolute top-2 -right-1 h-auto w-8 md:-right-3" />
+        <h2 className="py-5 text-center">
+          <span className="font-en text-en-h3 block uppercase">Sponsors</span>
+          <span className="text-ch-h4 block font-bold">鑽石級贊助商</span>
+        </h2>
+        <YellowStarIcon className="absolute bottom-8 -left-3 md:-left-8" />
+      </div>
+      <div className="section mb-5 gap-6 md:flex-row md:gap-14">
+        {COMPANIES.map((company) => (
+          <ExternalLink key={company.href} href={company.href} className="bg-n1 rounded-card p-3">
+            <Image alt={company.name} src={company.logo} />
+          </ExternalLink>
+        ))}
+      </div>
+    </section>
+  )
+}
